fix: DB tables and entity
database and relationship with the entity and user creation

+++ src/main/java/dev/clinican/dto/DoctorDto.java
@@ -8,6 +8,6 @@ public record DoctorDto(
         UUID userId,
         Long crm,
         String speciality,
-        Long phoneNumber
+        Integer phoneNumber
 ) {
 }
